@@ -13,20 +13,16 @@
 
 """Module houses class that implements ``PandasDataframePartitionManager``."""
 
-import numpy as np
+from distributed.client import default_client
 
 from modin.core.dataframe.pandas.partitioning.partition_manager import (
     PandasDataframePartitionManager,
 )
-from .axis_partition import (
+from .virtual_partition import (
     PandasOnDaskDataframeColumnPartition,
     PandasOnDaskDataframeRowPartition,
 )
 from .partition import PandasOnDaskDataframePartition
-from modin.error_message import ErrorMessage
-import pandas
-
-from distributed.client import default_client
 
 
 class PandasOnDaskDataframePartitionManager(PandasDataframePartitionManager):
@@ -38,93 +34,19 @@ class PandasOnDaskDataframePartitionManager(PandasDataframePartitionManager):
     _row_partition_class = PandasOnDaskDataframeRowPartition
 
     @classmethod
-    def get_indices(cls, axis, partitions, index_func):
+    def get_objects_from_partitions(cls, partitions):
         """
-        Get the internal indices stored in the partitions.
+        Get the objects wrapped by `partitions` in parallel.
 
         Parameters
         ----------
-        axis : {0, 1}
-            Axis to extract the labels over.
         partitions : np.ndarray
-            The array of partitions from which need to extract the labels.
-        index_func : callable
-            The function to be used to extract the indices.
+            NumPy array with ``PandasDataframePartition``-s.
 
         Returns
         -------
-        pandas.Index
-            A pandas Index object.
-
-        Notes
-        -----
-        These are the global indices of the object. This is mostly useful
-        when you have deleted rows/columns internally, but do not know
-        which ones were deleted.
+        list
+            The objects wrapped by `partitions`.
         """
         client = default_client()
-        ErrorMessage.catch_bugs_and_request_email(not callable(index_func))
-        func = cls.preprocess_func(index_func)
-        if axis == 0:
-            # We grab the first column of blocks and extract the indices
-            new_idx = (
-                [idx.apply(func).future for idx in partitions.T[0]]
-                if len(partitions.T)
-                else []
-            )
-        else:
-            new_idx = (
-                [idx.apply(func).future for idx in partitions[0]]
-                if len(partitions)
-                else []
-            )
-        new_idx = client.gather(new_idx)
-        return new_idx[0].append(new_idx[1:]) if len(new_idx) else new_idx
-
-    @classmethod
-    def broadcast_apply(cls, axis, apply_func, left, right, other_name="r"):
-        """
-        Broadcast the `right` partitions to `left` and apply `apply_func` function.
-
-        Parameters
-        ----------
-        axis : {0, 1}
-            Axis to apply and broadcast over.
-        apply_func : callable
-            Function to apply.
-        left : np.ndarray
-            NumPy array of left partitions.
-        right : np.ndarray
-            NumPy array of right partitions.
-        other_name : str, default: "r"
-            Name of key-value argument for `apply_func` that
-            is used to pass `right` to `apply_func`.
-
-        Returns
-        -------
-        np.ndarray
-            NumPy array of result partition objects.
-        """
-
-        def map_func(df, *others):
-            other = pandas.concat(others, axis=axis ^ 1)
-            return apply_func(df, **{other_name: other})
-
-        map_func = cls.preprocess_func(map_func)
-        rt_axis_parts = cls.axis_partition(right, axis ^ 1)
-        return np.array(
-            [
-                [
-                    part.apply(
-                        map_func,
-                        *(
-                            rt_axis_parts[col_idx].list_of_blocks
-                            if axis
-                            else rt_axis_parts[row_idx].list_of_blocks
-                        ),
-                    )
-                    for col_idx, part in enumerate(left[row_idx])
-                ]
-                for row_idx in range(len(left))
-            ]
-        )
+        return client.gather([partition.future for partition in partitions])
