@@ -14,7 +14,7 @@
 """Module houses `FWFDispatcher` class, that is used for reading of tables with fixed-width formatted lines."""
 
 import pandas
-from pandas._typing import FilePathOrBuffer
+from typing import Union, Sequence
 
 from modin.core.io.text.text_file_dispatcher import TextFileDispatcher
 
@@ -27,8 +27,10 @@ class FWFDispatcher(TextFileDispatcher):
     @classmethod
     def check_parameters_support(
         cls,
-        filepath_or_buffer: FilePathOrBuffer,
+        filepath_or_buffer,
         read_kwargs: dict,
+        skiprows_md: Union[Sequence, callable, int],
+        header_size: int,
     ):
         """
         Check support of parameters of `read_fwf` function.
@@ -39,6 +41,10 @@ class FWFDispatcher(TextFileDispatcher):
             `filepath_or_buffer` parameter of `read_fwf` function.
         read_kwargs : dict
             Parameters of `read_fwf` function.
+        skiprows_md : int, array or callable
+            `skiprows` parameter modified for easier handling by Modin.
+        header_size : int
+            Number of rows that are used by header.
 
         Returns
         -------
@@ -49,4 +55,6 @@ class FWFDispatcher(TextFileDispatcher):
             # If infer_nrows is a significant portion of the number of rows, pandas may be
             # faster.
             return False
-        return super().check_parameters_support(filepath_or_buffer, read_kwargs)
+        return super().check_parameters_support(
+            filepath_or_buffer, read_kwargs, skiprows_md, header_size
+        )
